+++ include/BFS_with_Grey.h
@@ -1,10 +1,10 @@
 #include "includes.h"
 
 template<typename Container, typename Adaptation_func>
-std::pair<typename Container::value_type, long int> BFS_Grey_algo(Container& using_field, const Adaptation_func& func, int N) {
+std::pair<typename Container::value_type, long double> BFS_Grey_algo(Container& using_field, const Adaptation_func& func, int N) {
     int position = rand() % using_field.size();
-    std::pair<typename Container::value_type, long int> max(using_field[position], func(using_field[position]));
-    std::pair<typename Container::value_type, long int> temp_pair=max;
+    std::pair<typename Container::value_type, long double> max(using_field[position], func(using_field[position]));
+    std::pair<typename Container::value_type, long double> temp_pair=max;
     for (int i = 0; i < N; i++) {
         std::cout <<"***** STEP "<<i<<" *****\n";
         std::cout << "Current max data: ";
